feat: update hook to take content and title as parameters
