+++ apps/web/src/components/workflow/preview/sms/SmsPreview.tsx
@@ -1,7 +1,10 @@
 import styled from '@emotion/styled';
 import { colors } from '@novu/design-system';
+import { useFormContext, useWatch } from 'react-hook-form';
+import { IForm } from '../../../../pages/templates/components/formTypes';
 import { useNavigateToStepEditor } from '../../../../pages/templates/hooks/useNavigateToStepEditor';
 import { usePreviewSmsTemplate } from '../../../../pages/templates/hooks/usePreviewSmsTemplate';
+import { useStepFormPath } from '../../../../pages/templates/hooks/useStepFormPath';
 import { useTemplateLocales } from '../../../../pages/templates/hooks/useTemplateLocales';
 import { LocaleSelect, MobileSimulator } from '../common';
 import { SmsBubble } from './SmsBubble';
@@ -25,7 +28,18 @@ const LocaleSelectStyled = styled(LocaleSelect)`
 
 export const SmsPreview = () => {
   const { navigateToStepEditor } = useNavigateToStepEditor();
-  const { selectedLocale, locales, areLocalesLoading, onLocaleChange } = useTemplateLocales();
+
+  const { control } = useFormContext<IForm>();
+  const path = useStepFormPath();
+  const templateContent = useWatch({
+    name: `${path}.template.content`,
+    control,
+  });
+
+  const { selectedLocale, locales, areLocalesLoading, onLocaleChange } = useTemplateLocales({
+    content: templateContent as string,
+  });
+
   const { isPreviewContentLoading, previewContent, templateContentError } = usePreviewSmsTemplate(selectedLocale);
 
   return (
