+++ apps/web/src/pages/templates/hooks/useTemplateLocales.ts
@@ -1,27 +1,28 @@
 import { useEffect, useState } from 'react';
-import { useFormContext, useWatch } from 'react-hook-form';
 
 import { useGetLocalesFromContent } from '../../../api/hooks';
 import { useAuthController, useDataRef } from '../../../hooks';
-import { IForm } from '../components/formTypes';
-import { useStepFormPath } from './useStepFormPath';
 
-export const useTemplateLocales = () => {
+export const useTemplateLocales = ({ content, title }: { content: string; title?: string }) => {
   const { organization } = useAuthController();
   const [selectedLocale, setSelectedLocale] = useState('');
-  const { control } = useFormContext<IForm>();
-  const path = useStepFormPath();
-  const templateContent = useWatch({
-    name: `${path}.template.content`,
-    control,
-  });
-  const previewData = useDataRef({ templateContent });
+
+  const previewData = useDataRef({ content, title });
 
   const { data: locales, isLoading: areLocalesLoading, getLocalesFromContent } = useGetLocalesFromContent();
 
   useEffect(() => {
+    let combinedContent = previewData.current.content;
+    /*
+     * combining title and content to get locales based upon variables in both title and content
+     * The api is not concerned about the content type, it will parse the given string and return the locales
+     */
+    if (previewData.current.title) {
+      combinedContent += ` ${previewData.current.title}`;
+    }
+
     getLocalesFromContent({
-      content: previewData.current.templateContent,
+      content: combinedContent,
     });
   }, [getLocalesFromContent, previewData]);
